requirements.txt
Pillow==8.4.0
protobuf==3.17.3
pytesseract==0.3.8
six==1.16.0
